atomizace user routeru
