+++ backend/routes/users/list.js
@@ -1,0 +1,19 @@
+/* eslint-disable */
+const express = require('express');
+const router = express.Router();
+const pool = require('../../config/database');
+const auth = require('../../middleware/auth');
+
+router.get('/', auth, async (req, res) => {
+  try {
+    const result = await pool.query(
+      'SELECT id, firstname, surname, email, role FROM users ORDER BY surname ASC'
+    );
+    res.json(result.rows);
+  } catch (error) {
+    console.error('Fetch users error:', error);
+    res.status(500).json({ error: 'Failed to fetch users' });
+  }
+});
+
+module.exports = router;
